+++ Sphinx/ChangeLog.rst
@@ -1,6 +1,8 @@
 ChangeLog
 =========
-* October 7, 2014
+* October 8, 2014
+
+  - Added information when pygame is not installed.
 
   - Corrected local filename bug in ``_load_media()``. (Thanks to Sergey Sorokin.)
   - Updated documentation.