+++ Sphinx/ChangeLog.rst
@@ -4,7 +4,7 @@ ChangeLog
 
   - Added information when pygame is not installed.
 
-  - Corrected local filename bug in ``_load_media()``. (Thanks to Sergey Sorokin.)
+  - **Corrected local filename bug in ``_load_media()``.** (Thanks to Sergey Sorokin.)
   - Updated documentation.
 
 * 01.08.00 --- October 4, 2014